Improve collision tolerance
The previous fix of just reducing bird size did not take into account
the fact that 0,0 is in the top-left corner so for the top (and left)
collisions, it did nothing.

The new fix adds a slight bit of overlap that's needed to be a real
collision, otherwise the user has the feeling they didn't really touch
anything yet still they got a game over.